--- pkg/registry/apis/identity/legacy/queries.go
@@ -1,22 +1,0 @@
-package legacy
-
-import (
-	"embed"
-	"fmt"
-	"text/template"
-)
-
-// Templates setup.
-var (
-	//go:embed *.sql
-	sqlTemplatesFS embed.FS
-
-	sqlTemplates = template.Must(template.New("sql").ParseFS(sqlTemplatesFS, `*.sql`))
-)
-
-func mustTemplate(filename string) *template.Template {
-	if t := sqlTemplates.Lookup(filename); t != nil {
-		return t
-	}
-	panic(fmt.Sprintf("template file not found: %s", filename))
-}
