+++ pkg/registry/apis/identity/legacy/testdata/mysql--display_query-display_ids_uids.sql
