+++ pkg/registry/apis/identity/legacy/testdata/postgres--display_query-display_uids.sql
