+++ pkg/registry/apis/identity/legacy/testdata/sqlite--display_query-display_uids.sql
